requirements.txt
fastapi
uvicorn
telethon
python-dotenv
sqlite-utils
python-multipart
aiofiles
